+++ lib/grizzly_series.dart
@@ -1,12 +1,15 @@
-// Copyright (c) 2017, SERAGUD. All rights reserved. Use of this source code
+// Copyright (c) 2017, Ravi Teja Gudapati. All rights reserved. Use of this source code
 // is governed by a BSD-style license that can be found in the LICENSE file.
 
-/// Support for doing something awesome.
+/// Provides
 ///
-/// More dartdocs go here.
+/// 1. [Array1D]
+/// 2. [Array2D]
+/// 3. [Series]
+/// 4. [DataFrame]
 library grizzly_series;
 
-export 'src/primitives/primitives.dart';
+export 'package:grizzly_primitives/grizzly_primitives.dart';
 export 'src/array/array.dart';
 export 'src/array2d/array2d.dart';
 export 'src/data_frame/data_frame.dart';